+++ autolog-spring/src/test/java/com/github/maximevw/autolog/test/MethodInOutAnnotatedTestClass.java
@@ -24,11 +24,11 @@ import com.github.maximevw.autolog.core.annotations.AutoLogMethodInOut;
 import com.github.maximevw.autolog.core.annotations.AutoLogMethodInput;
 import com.github.maximevw.autolog.core.annotations.AutoLogMethodOutput;
 import com.github.maximevw.autolog.core.logger.LogLevel;
-import com.github.maximevw.autolog.spring.aspects.AutoLogMethodInOutAspect;
+import com.github.maximevw.autolog.spring.aspects.AutoLogMethodInOutSpringAspect;
 
 /**
  * This class annotated with {@link AutoLogMethodInOut} contains methods specially designed to test
- * {@link AutoLogMethodInOutAspect}.
+ * {@link AutoLogMethodInOutSpringAspect}.
  */
 @AutoLogMethodInOut
 public class MethodInOutAnnotatedTestClass {